:lipstick:design: enhance feed post card UI 카드 UI 변경 카드 z index 변경에 따른 스크롤 버튼 z index 변경 보기 모드 삭제 카드 파일 위치 변경 Related to: #78
