+++ src/components/commons/ScrollToTop.tsx
@@ -22,7 +22,7 @@ const ScrollToTop = () => {
           position: 'fixed',
           bottom: 32,
           right: 32,
-          zIndex: 1,
+          zIndex: 3,
         }}
       >
         <Fab
